[SOC I/O] Fix GPI SMI handling
[Issue Description]
GPI SMI are not working.

[Resolution]
Fixed function to return 64-bit addresses.

Package/Module: OneSiliconPkg/Gpio

[Impacted Platform]
New platform

Hsd-es-id: 15017949926
Original commit date: Fri Jun 13 14:36:04 2025 +0200
Original commit hash: 48173fd1d38f4844db0b4a35424908333bb578ea

+++ OneSiliconPkg/Include/GpioV2Services.h
@@ -1111,8 +1111,8 @@ EFI_STATUS
   IN  UINT32             Index,
   OUT GPIOV2_PAD         *GpioPad,
   OUT UINT8              *GpiSmiBitOffset,
-  OUT UINT32             *GpiHostSwOwnRegAddress,
-  OUT UINT32             *GpiSmiStsRegAddress
+  OUT UINT64             *GpiHostSwOwnRegAddress,
+  OUT UINT64             *GpiSmiStsRegAddress
   );
 
 /**
